toggle button and from created

+++ Client/src/components/home.js
@@ -2,15 +2,22 @@ import React from 'react';
 
 
 function Home() {
-    return (
-        <div>
-            <h1>U Have Logged In Successfully</h1>
-            <button type="submit" onClick={(e)=>{
-                window.localStorage.clear();
-                window.location='/signup';
-            }}>Clear</button>
-        </div>
-    );
+    if (window.localStorage.getItem('token')) {
+        return (
+            <div>
+                <h1>U Have Logged In Successfully</h1>
+                <button type="submit" onClick={(e)=>{
+                    window.localStorage.clear();
+                    window.location='/';
+                }}>Log out</button>
+            </div>
+        );
+    }
+    else{
+        window.location = '/'
+    }
+   
+    
 }
 
 export default Home;
